first working z-mode version
adds a LOT of code, but should be more robust for the future, and even
though we’re essentially replacing one line of shell escaping via Regex
with ~100+, this should be the right way to do things.  Still not
entirely happy with the way we have to pull from buffer, but pending
reply from the git mailing list, lets stick with that for now.

The nice thing is we handle relative and multi paths WAY better then
scm_breeze now, woo.