start cleanup & rationalize for shortcode decoding
this section of the code is still not nearly as elegant as I would like,
and I would like to refactor the conditional logic/branching here, but
this change cleans up the syntax and formatting enough such that it will
be even possible to reason about such a refactor in the future.

in the meantime, being able to read is nice.