use goxc for cross compile
probably will need to tune this a bit more when we work towards an
actual release, but seems to be functioning well for now.

piggybacks on the version script to set binary version, if goxc is run
outside of rake task will just make a “snapshot” build which is fine
for now.