switch from nut to godep
Nut doesn’t seem to be actively maintained anymore, and hasn’t kept
pace with go1.6 vendor changes.

Long term, I believe glide will be a better option than godep, and we
can hopefully drop keeping the vendor/ directory in VCS, but for now
let’s stick with this methodology so as not to confuse downstream build
scripts etc.

As a note, `godep save ./…` complained at the mousetrap package not
being present — my assumption here is this was based on it only being
required by the windows path, which the local env didn’t require.
Manually `go get`ing it first resolved that issue for now.