change the way we handle @outside-repo
In the past, overriding dirs for the entire test suite was unreliable,
so we only did it for the few tests that absolutely needed to take
place outside of a git repository.

the way we did this in the past seems to have broken, but changing it
globally now seems to actually work, so deprecate the entire notion of
`@outside-repo` tagged tests, and move the Aruba working directory into
`/tmp`, so now all tests should be outside of a git tree unless it’s
one we created ourselves with the tests.