refactor(status): move process error handling up
This was originally written early in my Go programming days, and as a
CLI tool there are a lot of log.Fatals deep in the stack showing for
reporting fatal parsing conditions to the user. But, this makes writing
new tests a bit more difficult.

In preparation for starting to debug the process parsing bugs, extract
the error handling to be more library like.