fix(init): remove local from POSIX scripts
fixes shellcheck SC3043:
https://github.com/koalaman/shellcheck/wiki/SC3043

strictly speaking, local is not POSIX, but is supported in many shells,
including bash, ksh, dash, and BusyBox ash.

We could disable this rule, but perhaps safer to just do the effort
to try to be fully POSIX compliant for now, even though it makes
the script more difficult to read and reason about.