Remove duplicated fish flag from README
To match usage in CLI help:

> scmpuff init --help

Outputs the shell initialization script for scmpuff.

Initialize scmpuff by adding the following to your ~/.bash_profile or ~/.zshrc:

    eval "$(scmpuff init --shell=sh)"

For fish shell, add the following to ~/.config/fish/config.fish instead:

    scmpuff init --shell=fish | source

There are a number of flags to customize the shell integration.

Usage:
  scmpuff init [flags]

Flags:
  -a, --aliases                 Include short git aliases (default true)
  -h, --help                    help for init
  -s, --shell string[="fish"]   Output shell type: sh | bash | zsh | fish
  -w, --wrap                    Wrap standard git commands (default true)