refactor(status): move path calculations out of processing
This cleans up Process substantially and makes it better prepared for
a future modularized replacement, by removing path calculations and only
returning paths in a way that is closer to what is used by git itself.

However, it makes Renderer even messier, and is considered a partial
refactoring as there is a clear need to now rationalize and structure
how git repo root and user current working directory (cwd) are tracked.

Notably this also standardizes paths to always display in POSIX format,
even on Windows, as this appears to be what git prefers.  We may revisit
this in the future.