feat: new dumptool for repo debugging
Introduces a prototype dumptool, which bundles up a bunch of git repo
porcelain output into a zip file, to make it easy to collect samples
reproducible from end users on a repo that won't parse.

combined with the new golden file parse scaffolding introduced in the
previous commit, this has the potential to replace a huge chunk of
the Cucumber integration tests for status.

as a bonus, it can be used by developers with `-summary=false -porcelain`
to manually get porcelain sample files in various formats when working
on scmpuff.

dumptool has not been integrated into the scmpuff UI here, but to do
so in the future we would want to add something like
`scmpuff status --debug-dump` with appropriate user messaging to let
them know how to file an issue with the zip file.