refactor: adopt github.com/mroth/porcelain/statusv1 for status parsing
This is the minimum set of changes to slap my experimental git status
parsing module in to the existing code. This externalizes git status
parsing for separation of concerns

Additionally, it is likely we would want to switch to using porcelainv2
in the near future as enabled by github.com/mroth.porcelain, but a
phased approach is the plan here.